+++ simple_history/tests/tests/test_models.py
@@ -2847,7 +2847,7 @@ class HistoricForeignKeyTest(TestCase):
         # test querying directly from the history table and converting
         # to an instance, it should chase the foreign key properly
         # in this case if _as_of is not present we use the history_date
-        # https://github.com/jazzband/django-simple-history/issues/983
+        # https://github.com/django-commons/django-simple-history/issues/983
         pt1h = TestHistoricParticipanToHistoricOrganization.history.all()[0]
         pt1i = pt1h.instance
         self.assertEqual(pt1i.organization.name, "modified")
@@ -2974,7 +2974,7 @@ class HistoricOneToOneFieldTest(TestCase):
         # test querying directly from the history table and converting
         # to an instance, it should chase the foreign key properly
         # in this case if _as_of is not present we use the history_date
-        # https://github.com/jazzband/django-simple-history/issues/983
+        # https://github.com/django-commons/django-simple-history/issues/983
         pt1h = TestHistoricParticipanToHistoricOrganizationOneToOne.history.all()[0]
         pt1i = pt1h.instance
         self.assertEqual(pt1i.organization.name, "modified")