+++ CHANGES.rst
@@ -13,6 +13,8 @@ Unreleased
 - Dropped support for Python 3.8, which reached end-of-life on 2024-10-07 (gh-1421)
 - Added support for Django 5.1 (gh-1388)
 - Added pagination to ``SimpleHistoryAdmin`` (gh-1277)
+- Fixed issue with history button not working when viewing historical entries in the
+  admin (gh-527)
 
 3.7.0 (2024-05-29)
 ------------------