+++ CHANGES.rst
@@ -15,6 +15,7 @@ Unreleased
 - Added pagination to ``SimpleHistoryAdmin`` (gh-1277)
 - Fixed issue with history button not working when viewing historical entries in the
   admin (gh-527)
+- Fixed history_change_reason_field being shared across inherited models (gh-1433)
 
 3.7.0 (2024-05-29)
 ------------------